loginSteam -> Separado o login em um arquivo diferente da rota
